2023.02.01 11:14 Update

+++ Stack/0150. Evaluate Reverse Polish Notation.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
 	int evalRPN(vector<string>& tokens) {
-		stack<int> S;
+		stack<long long> S;
 		for (const string& token : tokens)
 		{
 			if (string{ "+-*/" }.find(token.back()) != string::npos)
 			{
-				int b = S.top(); S.pop();
-				int a = S.top(); S.pop();
+				long long b = S.top(); S.pop();
+				long long a = S.top(); S.pop();
 				if (token == "+")
 					S.push(a + b);
 				else if (token == "-")
@@ -18,7 +18,7 @@ public:
 					S.push(a / b);
 			}
 			else
-				S.push(atoi(token.c_str()));
+				S.push(atoll(token.c_str()));
 		}
 		return S.top();
 	}
